`BaseSampler` should not change the dimensions of the sampled tensors.
While writing an experiment for the MNIST dataset, I discovered that
`BaseSampler` squeezes the sampled tensors. This caused a bug in my
experiment b/c the MNIST dataset is black and white and therefore only
has one channel, i.e. the sampled tensor was of dimension
`(mini-batch-size, 1, width, height)`, but it got squeezed by
`BaseSampler` to be `(mini-batch-size, width, height)`. This took me a
bit to debug b/c I didn't expect the sampler to change the dimensions.
This commit removes that code. I think it's more natural that the
dimensions are handled before the tensors are handed over to the
sampler.